fix call to showwarning in generateCoefficients.py
playing around with showwarning in other python scrips,
the filename argument is supposed to be the name of the
source code file raising the warning.  The lineno argument
is required and represents the line number in the source
code that is raising the warning.  The file kwarg is the
file handle to which the warning is written.